New stuff coming up!

+++ front-end/src/app/pages/sitecp/sub-pages/shop/items/item/item.component.ts
@@ -10,6 +10,7 @@ import { TitleTab } from 'shared/app-views/title/title.model';
 import { SlimSubscription } from '../../../users/subscriptions/subscriptions.model';
 import { ItemService } from '../../services/item.service';
 import { NotificationService } from 'core/services/notification/notification.service';
+import { DialogService } from 'core/services/dialog/dialog.service';
 
 @Component({
     selector: 'app-sitecp-shop-item',
@@ -28,6 +29,7 @@ export class ItemComponent extends Page implements OnDestroy {
         private _router: Router,
         private _service: ItemService,
         private _notificationService: NotificationService,
+        private _dialogService: DialogService,
         elementRef: ElementRef,
         breadcrumbService: BreadcrumbService,
         activatedRoute: ActivatedRoute
@@ -136,9 +138,16 @@ export class ItemComponent extends Page implements OnDestroy {
     }
 
     private onDelete (): void {
-        this._service.delete(this._data.shopItemId).subscribe(() => {
-            this._router.navigateByUrl(SHOP_ITEMS_BREADCRUMB_ITEMS.url);
-        }, this._notificationService.failureNotification.bind(this._notificationService));
+        this._dialogService.confirm({
+            title: 'Are you sure?',
+            content: 'Are you sure that you wanna delete this item?',
+            callback: () => {
+                this._dialogService.closeDialog();
+                this._service.delete(this._data.shopItemId).subscribe(() => {
+                    this._router.navigateByUrl(SHOP_ITEMS_BREADCRUMB_ITEMS.url);
+                }, this._notificationService.failureNotification.bind(this._notificationService));
+            }
+        });
     }
 
     private setTabs (): void {
